fix(utils): seperated all utils functions
