+++ packages/core/src/types/Attributes.ts
@@ -1,0 +1,1 @@
+export type Attributes = { [key: string]: any };
