+++ packages/core/src/utils/getNextSibling.ts
@@ -1,0 +1,3 @@
+const getNextSibling = (element: HTMLElement) => element.nextElementSibling as HTMLElement;
+
+export default getNextSibling;
