+++ packages/core/src/utils/getPreviousSibling.ts
@@ -1,0 +1,3 @@
+const getPreviousSibling = (element: HTMLElement) => element.previousElementSibling as HTMLElement;
+
+export default getPreviousSibling;
